# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_muchlist_moneymagnet_business_spend_model.NewSpend:
    properties:
      category_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FXX
        type: string
      date:
        example: "2022-09-10T17:03:15.091267+08:00"
        type: string
      id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FZZ
        type: string
      is_income:
        example: false
        type: boolean
      name:
        example: Makan siang
        type: string
      pocket_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FYY
        type: string
      price:
        example: 50000
        type: integer
      type:
        example: 2
        type: integer
    type: object
  github_com_muchlist_moneymagnet_business_spend_model.SpendResp:
    properties:
      balance_snapshoot:
        example: 0
        type: integer
      category_icon:
        example: 1
        type: integer
      category_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FXX
        type: string
      category_name:
        example: food
        type: string
      created_at:
        example: "2022-09-10T17:03:15.091267+08:00"
        type: string
      date:
        example: "2022-09-10T17:03:15.091267+08:00"
        type: string
      id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FZZ
        type: string
      is_income:
        example: false
        type: boolean
      name:
        example: Makan siang
        type: string
      pocket_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FYY
        type: string
      pocket_name:
        example: main pocket
        type: string
      price:
        example: 50000
        type: integer
      type:
        example: 2
        type: integer
      updated_at:
        example: "2022-09-10T17:03:15.091267+08:00"
        type: string
      user_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FNN
        type: string
      user_name:
        example: Muchlis
        type: string
      version:
        example: 1
        type: integer
    type: object
  github_com_muchlist_moneymagnet_business_spend_model.TransferSpend:
    properties:
      date:
        example: "2022-09-10T17:03:15.091267+08:00"
        type: string
      pocket_id_from:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FYY
        type: string
      pocket_id_to:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FXX
        type: string
      price:
        example: 50000
        type: integer
    type: object
  github_com_muchlist_moneymagnet_business_spend_model.UpdateSpend:
    properties:
      category_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FXX
        type: string
      date:
        example: "2022-09-10T17:03:15.091267+08:00"
        type: string
      is_income:
        example: false
        type: boolean
      name:
        example: Makan siang
        type: string
      price:
        example: 50000
        type: integer
      type:
        example: 2
        type: integer
    type: object
  github_com_muchlist_moneymagnet_pkg_data.Metadata:
    properties:
      current_page:
        example: 1
        type: integer
      first_page:
        example: 1
        type: integer
      last_page:
        example: 1
        type: integer
      page_size:
        example: 50
        type: integer
      total_records:
        example: 1
        type: integer
    type: object
  misc.Response500Err:
    properties:
      error:
        example: 'name func: sub func: cause of error'
        type: string
    type: object
  misc.ResponseErr:
    properties:
      error:
        example: example error message
        type: string
      error_field:
        additionalProperties:
          type: string
        example:
          example_field: example_field is a required field
        type: object
    type: object
  misc.ResponseMessage:
    properties:
      data:
        example: do thing success
        type: string
    type: object
  misc.ResponseSuccess:
    properties:
      data: {}
    type: object
  misc.ResponseSuccessList:
    properties:
      data:
        items: {}
        type: array
      meta_data:
        $ref: '#/definitions/github_com_muchlist_moneymagnet_pkg_data.Metadata'
    type: object
host: localhost
info:
  contact:
    email: whois.muchlis@gmail.com
    name: Muchlis
    url: https://muchlis.dev
  description: this is server for money magnet application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Money Magnet API
  version: "1.0"
paths:
...
  /spends:
    get:
      consumes:
      - application/json
      description: Find spend
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: page-size
        in: query
        name: page_size
        type: integer
      - description: sort
        in: query
        name: sort
        type: string
      - description: user
        in: query
        name: user
        type: string
      - description: category
        in: query
        name: category
        type: string
      - description: is_income
        in: query
        name: is_income
        type: boolean
      - description: type
        in: query
        name: type
        type: string
      - description: date_start
        in: query
        name: date_start
        type: integer
      - description: date_end
        in: query
        name: date_end
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/misc.ResponseSuccessList'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_muchlist_moneymagnet_business_spend_model.SpendResp'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/misc.ResponseErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/misc.Response500Err'
      summary: Find Spend
      tags:
      - Spend
    post:
      consumes:
      - application/json
      description: Create spend
      parameters:
      - description: Request Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/github_com_muchlist_moneymagnet_business_spend_model.NewSpend'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/misc.ResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/github_com_muchlist_moneymagnet_business_spend_model.SpendResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/misc.ResponseErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/misc.Response500Err'
      summary: Create Spend
      tags:
      - Spend
  /spends/{spend_id}:
    get:
      consumes:
      - application/json
      description: Get spend detail by ID
      parameters:
      - description: spend_id
        in: path
        name: spend_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/misc.ResponseSuccessList'
            - properties:
                data:
                  $ref: '#/definitions/github_com_muchlist_moneymagnet_business_spend_model.SpendResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/misc.ResponseErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/misc.Response500Err'
      summary: Get Spend Detail
      tags:
      - Spend
    patch:
      consumes:
      - application/json
      description: Update spend
      parameters:
      - description: spend_id
        in: path
        name: spend_id
        required: true
        type: string
      - description: Request Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/github_com_muchlist_moneymagnet_business_spend_model.UpdateSpend'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/misc.ResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/github_com_muchlist_moneymagnet_business_spend_model.SpendResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/misc.ResponseErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/misc.Response500Err'
      summary: Update Spend
      tags:
      - Spend
  /spends/sync/{spend_id}:
    post:
      consumes:
      - application/json
      description: Sync spend to update pocket balance
      parameters:
      - description: spend_id
        in: path
        name: spend_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.ResponseMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/misc.ResponseErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/misc.Response500Err'
      summary: Sync Spend Balance
      tags:
      - Spend
  /spends/transfer:
    post:
      consumes:
      - application/json
      description: Tramsfer
      parameters:
      - description: Request Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/github_com_muchlist_moneymagnet_business_spend_model.TransferSpend'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.ResponseMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/misc.ResponseErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/misc.Response500Err'
      summary: Transfer
      tags:
      - Spend
swagger: "2.0"